Revert "cherry-pick:  from odh incubation to 2.6"
